+++ index.d.ts
@@ -15,6 +15,7 @@ import { ChainLib, ChainConfig } from './typings/chain';
 import { MaximusChainLib } from './typings/chains/maximus';
 import { OsmiumChainLib } from './typings/chains/osmium';
 import { FilopowChainLib } from './typings/chains/filopow';
+import { FewbitChainLib } from './typings/chains/fewbit';
 
 /** Library version, e.g. `"v3.0.2"`. */
 export const version: string;
@@ -49,6 +50,7 @@ export function createHashRegistry(): import('./typings/chain').HashRegistry;
 export function create(name: 'maximus'): MaximusChainLib;
 export function create(name: 'osmium'): OsmiumChainLib;
 export function create(name: 'filopow'): FilopowChainLib;
+export function create(name: 'fewbit'): FewbitChainLib;
 export function create(name: string): ChainLib;
 
 /**